requirements.txt
requests
python-dotenv
tweepy
pandas
google-cloud-storage